Remove IP2GEO tool due to obsolescence
The maplotlib used by the tool has been depreciated for over a year. Replaced by "cartographer" I think, and isn't supported.

This tool was originally added to just illustrate a point of clearnet node "traceability" now done better with web based tools and this project added a lot of bloat, and prone to install errors.

This also allows for standard php to be used, not version specific as before due to upload config for the ip2geo database.