Mass changes for prep to PiNodeXMR v5 (Part 1)
Years of development have resulted in poor file name choices, and a mass of files in user pinodexmr home dir.
Re-labeling, organising and increase robustness of newer procedures and use of systemd to create intuitive project construction. Aim to make this easier for user to naigate and hopefully contribute to.