Allows PiNodeXMR disk images to auto-expand file system on boot.
Raspberry Pi Ubuntu images always auto-expand.
Armbian images require the command 'sudo systemctl start armbian-resize-filesystem' to grow the file system to the extent of the card the image is on.

To handle this, PiNodeXMR disk images have the "bootstatus" vale set to -1 on a disk image only. This triggers the command on login only once.